@@ -1,89 +1,52 @@
 {
-  "name": "comic-universe",
-  "version": "2.0.0-alpha.6",
-  "description": "Open Source App made using Electron and React for downloading and reading comics from web sources provided by the user. This app was created for study purposes only and should never be sold or used for anything commercial",
-  "main": "./out/main/index.js",
-  "author": "Pablo Souza",
-  "repository": "https://github.com/pablovsouza/comic-universe",
-  "license": "MIT",
+  "name": "comic-universe-tauri",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "format": "prettier --write .",
-    "lint": "eslint . --ext .js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix",
-    "typecheck:node": "tsc --noEmit -p tsconfig.node.json --composite false",
-    "typecheck:web": "tsc --noEmit -p tsconfig.web.json --composite false",
-    "typecheck": "npm run typecheck:node && npm run typecheck:web",
-    "start": "electron-vite preview",
-    "dev": "electron-vite dev",
-    "build": "npm run typecheck && electron-vite build",
-    "prisma:generate": "npx prisma generate",
-    "postinstall": "electron-builder install-app-deps && npx prisma generate",
-    "generate-icons": "electron-icon-builder --input=./resources/icon.png --output=build --flatten",
-    "build:win": "npm run prisma:generate && npm run build && electron-builder --win --config",
-    "build:linux-x64": "npm run prisma:generate && electron-vite build && electron-builder --linux --config",
-    "build:linux-arm": "npm run prisma:generate && electron-vite build && electron-builder --linux --arm64 --config",
-    "build:mac": "npm run prisma:generate && electron-vite build && electron-builder --mac --universal --config"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "tauri": "tauri"
   },
   "dependencies": {
-    "@electron-toolkit/preload": "^3.0.1",
-    "@electron-toolkit/utils": "^3.0.0",
-    "@electron/asar": "^3.2.7",
     "@formkit/auto-animate": "^0.8.2",
-    "@jridgewell/resolve-uri": "^3.1.2",
-    "@prisma/client": "^5.21.1",
     "@tanstack/react-query": "^5.55.1",
-    "axios": "^1.7.4",
+    "@tauri-apps/api": "^2.8.0",
+    "@tauri-apps/plugin-dialog": "^2.4.0",
+    "@tauri-apps/plugin-fs": "^2.4.2",
+    "@tauri-apps/plugin-http": "^2.5.2",
+    "@tauri-apps/plugin-opener": "^2.5.0",
+    "@tauri-apps/plugin-shell": "^2.3.1",
+    "axios": "^1.11.0",
     "classnames": "^2.3.2",
-    "cors": "^2.8.5",
-    "electron-updater": "^6.1.4",
-    "express": "^4.21.0",
+    "html-react-parser": "^5.2.3",
     "lodash.debounce": "^4.0.8",
-    "pacote": "^18.0.6",
-    "prisma-cli-extension": "^1.0.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-html-parser": "^2.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-paginate": "^8.2.0",
     "react-reverse-portal": "^2.1.2",
-    "react-router-dom": "^6.18.0",
+    "react-router-dom": "^6.30.1",
     "react-select": "^5.7.7",
     "react-transition-group": "^4.4.5",
     "slugify": "^1.6.6",
     "tw-colors": "^3.3.1",
-    "uuid": "^10.0.0",
-    "zustand": "^4.4.6"
+    "uuid": "^11.1.0",
+    "zustand": "^4.5.7"
   },
   "devDependencies": {
-    "@electron-toolkit/tsconfig": "^1.0.1",
-    "@electron/notarize": "^2.1.0",
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
+    "@tauri-apps/cli": "^2.8.4",
     "@types/lodash.debounce": "^4.0.8",
-    "@types/node": "22.5.4",
-    "@types/pacote": "^11.1.8",
-    "@types/react": "^18.2.35",
-    "@types/react-dom": "18.3.0",
-    "@types/react-html-parser": "^2.0.4",
+    "@types/node": "^24.3.1",
+    "@types/react": "^18.3.24",
+    "@types/react-dom": "^18.3.7",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.0.1",
-    "@typescript-eslint/parser": "^8.0.1",
-    "@vitejs/plugin-react": "^4.1.1",
+    "@vitejs/plugin-react": "^5.0.2",
     "autoprefixer": "^10.4.20",
-    "electron": "^33.0.2",
-    "electron-builder": "^25.1.8",
-    "electron-icon-builder": "^2.0.1",
-    "electron-vite": "^2.3.0",
-    "eslint": "^9.8.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-react": "^7.33.2",
     "postcss": "^8.4.41",
-    "prettier": "^3.0.3",
-    "prisma": "^5.21.1",
-    "sass-embedded": "^1.77.8",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^5.2.2",
-    "vite": "^5.4.0",
-    "vite-plugin-svgr": "^4.2.0"
+    "tailwindcss": "^3.4.17",
+    "typescript": "^5.9.2",
+    "vite": "^7.0.4"
   },
-  "packageManager": "yarn@4.5.1"
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
