@@ -13,8 +13,8 @@ const addTransparency = (color, transparency) => {
 }
 
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ['./src/renderer/index.html', './src/renderer/**/*.{js,ts,jsx,tsx}'],
+export default {
+  content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
   theme: {
     extend: {
       fontFamily: {
